+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Admin page uses `have-token` endpoint to determine if Google account is connected
+
 ## [0.0.11] - 2020-08-27
 
 ### Changed
@@ -29,13 +33,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [0.0.8] - 2020-08-25
 
 ### Added
- 
+
 - Outbound access
 
 ## [0.0.7] - 2020-08-25
 
 ### Added
- 
+
 - Added logging
 
 ## [0.0.6] - 2020-08-25